fix(channel-search): decode Go html.EscapeString entities (not Lucene forms)
Round-5 CR (yujiawei P1-1, Jerry-Xin blocker 1, Octo-Q P1) on head 3d360e9e:
decodeServerEscapedHighlight's table hand-wrote the Lucene
SimpleHTMLEncoder entity forms (&quot; / &#x27; / &#x2F;), but the paired
server PR Mininglamp-OSS/octo-server#834 uses Go's html.EscapeString
instead (asserted by TestBuildSearchFilesHighlight_NoRequestEncoder /
TestBuildSearchAllHighlight_NoRequestEncoder), which emits decimal forms
&#34; and &#39; and never escapes slash. Result: apostrophes and
double-quotes in file names and body snippets — extremely common,
verified against Tika-extracted English bodies — rendered as literal
&#39;s / &#34; entities in the UI once the paired server shipped.

Three changes:

  1. Fix the decode table: &#34; / &#39; / &lt; / &gt; / &amp; (last),
     drop &quot; / &#x27; / &#x2F; entirely. Doc comment rewritten to
     name Go html.EscapeString as the contract and point at
     TestEscapeHighlightFragment (the server test whose want-strings are
     now the source of client fixtures too).

  2. Move decodeHighlight.ts from Components/ChannelSearch/ to Service/
     as decodeServerEscapedHighlight.ts. Consumed by SearchResultMapper
     (a Service-layer module), so it belongs on the same side of the
     component boundary. This also cuts the Service → Components import
     back-edge that at head 600f721e transitively pulled the App tree
     (react-virtuoso, etc.) into vitest and broke four unrelated suites
     during collection. snippetContent.tsx keeps a re-export so the
     public API remains stable.

  3. Test fixtures replaced with strings lifted verbatim from the server's
     own tests (TestEscapeHighlightFragment, TestSingleFileHit_HostileContentEscaped
     in modules/messages_search/search_files_test.go), so the two suites
     can only disagree if someone changes the server on purpose. The
     four suites broken at head 600f721e still load and pass, guarding
     the layering fix.
